change VotePanel to VoteModal
